Don't crash when deleting a non-existent record
This shouldn't even happen, but it apparently happened to a user. This
commit changes it so it logs a warning instead of crashing.

+++ app/src/main/java/com/etesync/syncadapter/syncadapter/CalendarSyncManager.java
@@ -104,8 +104,12 @@ public class CalendarSyncManager extends SyncManager {
         if (cEntry.isAction(SyncEntry.Actions.ADD) || cEntry.isAction(SyncEntry.Actions.CHANGE)) {
             processEvent(event, local);
         } else {
-            App.log.info("Removing local record #" + local.getId() + " which has been deleted on the server");
-            local.delete();
+            if (local != null) {
+                App.log.info("Removing local record #" + local.getId() + " which has been deleted on the server");
+                local.delete();
+            } else {
+                App.log.warning("Tried deleting a non-existent record: " + event.uid);
+            }
         }
     }
